Fix CI problem.
