+++ test/config-test/src/test/java/com/alibaba/nacos/test/config/NacosConfigServiceNoComTlsGrpcClient_CITCase.java
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-
 package com.alibaba.nacos.test.config;
 
 import com.alibaba.nacos.Nacos;
@@ -26,7 +25,7 @@ import com.alibaba.nacos.client.config.listener.impl.AbstractConfigChangeListene
 import com.alibaba.nacos.common.remote.client.RpcConstants;
 import com.alibaba.nacos.core.remote.RpcServerTlsConfig;
 import com.alibaba.nacos.test.base.ConfigCleanUtils;
-import org.junit.After;
+import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -51,46 +50,40 @@ import java.util.concurrent.atomic.AtomicInteger;
         properties = {
                 "nacos.standalone=true",
                 RpcServerTlsConfig.PREFIX+".enableTls=true",
-                RpcServerTlsConfig.PREFIX+".mutualAuthEnable=true",
                 RpcServerTlsConfig.PREFIX+".compatibility=false",
                 RpcServerTlsConfig.PREFIX+".certChainFile=test-server-cert.pem",
-                RpcServerTlsConfig.PREFIX+".certPrivateKey=test-server-key.pem",
-                RpcServerTlsConfig.PREFIX+".trustCollectionCertFile=test-ca-cert.pem",
-
-        },
+                RpcServerTlsConfig.PREFIX+".certPrivateKey=test-server-key.pem"},
         webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-public class NacosConfigV2MutualAuthTest {
-
+public class NacosConfigServiceNoComTlsGrpcClient_CITCase {
 
     public static AtomicInteger increment = new AtomicInteger(100);
 
     @BeforeClass
-    public static   void beforeClass() throws IOException {
-        ConfigCleanUtils.changeToNewTestNacosHome(NacosConfigV2MutualAuthTest.class.getSimpleName());
+    public static void beforeClass() throws IOException {
+        ConfigCleanUtils.changeToNewTestNacosHome(NacosConfigServiceNoComTlsGrpcClient_CITCase.class.getSimpleName());
 
     }
 
-    @After
-    public  void cleanClientCache() throws Exception {
+    @BeforeClass
+    @AfterClass
+    public static void cleanClientCache() throws Exception {
         ConfigCleanUtils.cleanClientCache();
     }
 
     @Test
-    public void test_d_MutualAuth() throws Exception {
-        Properties propertiesfalse = new Properties();
-        propertiesfalse.put(RpcConstants.RPC_CLIENT_TLS_ENABLE, "true");
-        propertiesfalse.put(RpcConstants.RPC_CLIENT_MUTUAL_AUTH,"true");
-        propertiesfalse.put(RpcConstants.RPC_CLIENT_TLS_CERT_KEY,"test-client-key.pem");
-        propertiesfalse.put(RpcConstants.RPC_CLIENT_TLS_TRUST_COLLECTION_CHAIN_PATH,"test-ca-cert.pem");
-        propertiesfalse.put(RpcConstants.RPC_CLIENT_TLS_CERT_CHAIN_PATH,"test-client-cert.pem");
-        propertiesfalse.put("serverAddr", "127.0.0.1");
-        ConfigService configServiceFalse = new NacosConfigService(propertiesfalse);
+    public void test_e_TlsServerAndTlsClient() throws Exception {
+        Properties properties = new Properties();
+        properties.put(RpcConstants.RPC_CLIENT_TLS_ENABLE, "true");
+        properties.put(RpcConstants.RPC_CLIENT_TLS_PROVIDER, "openssl");
+        properties.put(RpcConstants.RPC_CLIENT_TLS_TRUST_COLLECTION_CHAIN_PATH, "test-ca-cert.pem");
+        properties.put("serverAddr", "127.0.0.1");
+        ConfigService configService = new NacosConfigService(properties);
+        String content = UUID.randomUUID().toString();
         String dataId = "test-group" + increment.getAndIncrement();
         String groupId = "test-data" + increment.getAndIncrement();
-        String content = UUID.randomUUID().toString();
-        boolean res = configServiceFalse.publishConfig(dataId, groupId, content);
-        CountDownLatch latch2 = new CountDownLatch(1);
-        configServiceFalse.addListener(dataId, groupId, new AbstractConfigChangeListener() {
+        boolean b = configService.publishConfig("test-group" + increment.getAndIncrement(), "test-data" + increment.getAndIncrement(), content);
+        CountDownLatch latch = new CountDownLatch(1);
+        configService.addListener(dataId, groupId, new AbstractConfigChangeListener() {
             @Override
             public void receiveConfigChange(ConfigChangeEvent event) {
                 ConfigChangeItem cci = event.getChangeItem("content");
@@ -98,20 +91,17 @@ public class NacosConfigV2MutualAuthTest {
                 if (!content.equals(cci.getNewValue())) {
                     return;
                 }
-                latch2.countDown();
+                latch.countDown();
             }
         });
-        latch2.await(5, TimeUnit.SECONDS);
-        Assert.assertTrue(res);
+        latch.await(5, TimeUnit.SECONDS);
+        Assert.assertTrue(b);
     }
 
     @Test
-    public void test_d_MutualAuthButClientNot() throws Exception {
-
+    public void test_e_TlsServerAndPlainClient()  throws Exception {
         Properties propertiesfalse = new Properties();
-        propertiesfalse.put(RpcConstants.RPC_CLIENT_TLS_ENABLE, "true");
-        propertiesfalse.put(RpcConstants.RPC_CLIENT_TLS_TRUST_COLLECTION_CHAIN_PATH,"test-client-cert.pem");
-
+        propertiesfalse.put(RpcConstants.RPC_CLIENT_TLS_ENABLE, "false");
         propertiesfalse.put("serverAddr", "127.0.0.1");
         ConfigService configServiceFalse = new NacosConfigService(propertiesfalse);
         String dataId = "test-group" + increment.getAndIncrement();
